agents/requirements.txt
# LangChain and AI/LLM
langchain
langchain-core
langchain-groq
langchain-community
langchain-huggingface
langchain_astradb
langchain-postgres
langgraph
groq
instructor

# Vector Databases and Embeddings
chromadb
pgvector
sentence-transformers

# Database and Storage
psycopg2-binary
sqlalchemy
cassio

# FastAPI and Web Framework
fastapi
uvicorn
slowapi
pydantic
pydantic-settings
python-dotenv
python-multipart

# HTTP Clients and Networking
httpx
requests

# Computer Vision and Image Processing
opencv-python
deepface
pillow
mediapipe

# Audio Processing
librosa
sounddevice

# Data Science and Machine Learning
numpy
pandas
scikit-learn
scipy==1.10.1
torch
tf-keras
altair

# External APIs and Data Sources
tavily-python
wikipedia
arxiv
youtube_search
youtube-transcript-api
googletrans

# Development Tools
pytest
black
isort
rich
typing_extensions